Update patch set 3
Patch Set 3:

If I understand this right, there are no tests, because the templates are not bundled but need to be in the same folder where the server is run. Wouldn't it be better to bundle the templates in the resources? Then starting a jar would work wherever it needs to work, making deployment significantly easier. Also it would at least be easier to test.

But: Are the templates really necessary? The client javascript could very well be served by Kalamar and just be loaded in the export.html. And the export.html could be directly in the Java source code ...

Patch-set: 3
Reviewer: Gerrit User 19 <19@626241b7-b7d0-4197-b843-1ee23887649c>